models/token: Inline `ApiToken::insert_with_scopes()` fn
