+++ src/models.rs
@@ -14,7 +14,7 @@ pub use self::krate::{Crate, CrateName, NewCrate, RecentCrateDownloads};
 pub use self::owner::{CrateOwner, Owner, OwnerKind};
 pub use self::rights::Rights;
 pub use self::team::{NewTeam, Team};
-pub use self::token::{ApiToken, CreatedApiToken};
+pub use self::token::ApiToken;
 pub use self::user::{NewUser, User};
 pub use self::version::{NewVersion, TopVersions, Version};
