+++ src/tests/util.rs
@@ -19,7 +19,7 @@
 //! `MockCookieUser` and `MockTokenUser` provide an `as_model` function which returns a reference
 //! to the underlying database model value (`User` and `ApiToken` respectively).
 
-use crate::models::{ApiToken, CreatedApiToken, User};
+use crate::models::{ApiToken, User};
 use crate::tests::{
     CategoryListResponse, CategoryResponse, CrateList, CrateResponse, GoodCrate, OwnerResp,
     OwnersResponse, VersionResponse,
@@ -28,7 +28,7 @@ use std::future::Future;
 
 use http::{Method, Request};
 
-use crate::models::token::{CrateScope, EndpointScope};
+use crate::models::token::{CrateScope, EndpointScope, NewApiToken};
 use crate::util::token::PlainToken;
 use axum::body::{Body, Bytes};
 use axum::extract::connect_info::MockConnectInfo;
@@ -50,6 +50,7 @@ mod mock_request;
 mod response;
 mod test_app;
 
+use crate::controllers::token::CreatedApiToken;
 use mock_request::MockRequest;
 pub use mock_request::MockRequestExt;
 pub use response::Response;
@@ -320,16 +321,21 @@ impl MockCookieUser {
     ) -> MockTokenUser {
         let mut conn = self.app().db_conn().await;
 
-        let token = ApiToken::insert_with_scopes(
-            &mut conn,
-            self.user.id,
-            name,
-            crate_scopes,
-            endpoint_scopes,
-            expired_at,
-        )
-        .await
-        .unwrap();
+        let plaintext = PlainToken::generate();
+
+        let new_token = NewApiToken::builder()
+            .user_id(self.user.id)
+            .name(name)
+            .token(plaintext.hashed())
+            .maybe_crate_scopes(crate_scopes)
+            .maybe_endpoint_scopes(endpoint_scopes)
+            .maybe_expired_at(expired_at)
+            .build();
+
+        let token = CreatedApiToken {
+            plaintext,
+            model: new_token.insert(&mut conn).await.unwrap(),
+        };
 
         MockTokenUser {
             app: self.app.clone(),
